Added timing in tests. Forward to graphic tests

+++ tester/tester.sh
@@ -61,21 +61,24 @@ for root_lib_dir  in $ROOT_LIB_DIR  ; do
 	lib_dir="$root_lib_dir/$MF_SUFFIX/$LIBOUT_SUFFIX/"
 	if [ -d $make_dir ]; then
 		out_dir="$root_lib_dir/$MF_SUFFIX/$LIBOUT_SUFFIX"
+		func_name="it_build_$(basename $root_lib_dir)"
+		echo "$func_name() {" >> $test_file
 		if [ -f $make_dir/CMakeLists.txt ]; then
-			echo "rm -rf $PWD/build_dir" >> $test_file
-			echo "mkdir -p $PWD/build_dir" >> $test_file
-			echo "pushd $PWD/build_dir" >> $test_file
-			echo "cmake '$CALL_PWD/$make_dir'" >> $test_file
-			echo "make" >> $test_file
-			echo "mkdir -p '$CALL_PWD/$out_dir'" >> $test_file
-			echo "cp -f './$LIBOUT_SUFFIX'/* '$CALL_PWD/$out_dir/'" >> $test_file
-			echo "rm -rf '$PWD/build_dir'" >> $test_file
-			echo "popd" >> $test_file
+			echo "	rm -rf $PWD/build_dir" >> $test_file
+			echo "	mkdir -p $PWD/build_dir" >> $test_file
+			echo "	pushd $PWD/build_dir" >> $test_file
+			echo "	cmake '$make_dir'" >> $test_file
+			echo "	make" >> $test_file
+			echo "	mkdir -p '$out_dir/'" >> $test_file
+			echo "	cp -f './$LIBOUT_SUFFIX'/* '$out_dir/'" >> $test_file
+			echo "	rm -rf '$PWD/build_dir'" >> $test_file
+			echo "	popd" >> $test_file
 		else
-			echo "pushd '$make_dir'" >> $test_file
-			echo "make" >> $test_file
-			echo "popd" >> $test_file
+			echo "	pushd '$make_dir'" >> $test_file
+			echo "	make" >> $test_file
+			echo "	popd" >> $test_file
 		fi
+		echo "}" >> $test_file
 		echo "" >> $test_file
 		for root_test_dir in $ROOT_TEST_DIR ; do
 			for test in $root_test_dir/$MF_SUFFIX/$TEST_SUFFIX/*.c ; do
@@ -88,11 +91,17 @@ for root_lib_dir  in $ROOT_LIB_DIR  ; do
 				echo "$func_name() {" >> $test_file
 				echo "    gcc $CFLAGS -I'$PWD/../include' -c -o '$test_object_name' $LDFLAGS '$test'" >> $test_file
 				echo "    g++ $CFLAGS -g -o '$test_out_name' '$test_object_name' $LDFLAGS -L'$out_dir'" >> $test_file
-				echo "    set -x" >> $test_file
-				echo "    $PREC '$test_out_name' '$PWD/small.txt' '$PWD/out.txt'" >> $test_file
-				echo "    $PREC '$test_out_name' '$PWD/medium.txt' '$PWD/out.txt'" >> $test_file
-				echo "    $PREC '$test_out_name' '$PWD/gpl.txt' '$PWD/out.txt'" >> $test_file
-				echo "    set +x" >> $test_file
+				echo "    (>&3 echo '$(basename $root_lib_dir) $(basename $root_test_dir) $(basename $test .c)')" >> $test_file
+				echo '    resarr[0]=' >> $test_file
+				echo '    for i in `seq 0 9`; do' >> $test_file
+				echo "        rm -rf ./times" >> $test_file
+				echo '        start=$(date +"%s.%N")' >> $test_file
+				echo "        $PREC '$test_out_name' '$PWD/gpl.txt' '$PWD/out.txt'" >> $test_file
+				echo '        end=$(date +"%s.%N")' >> $test_file
+				echo '        resarr[$i]=$(bc <<< "$end-$start")' >> $test_file
+				echo "    done" >> $test_file
+				echo '    (>&3 echo ${resarr[*]})' >> $test_file
+				echo '    (>&3 echo "")' >> $test_file
 				echo "}" >> $test_file
 				echo "" >> $test_file
 			fi
@@ -108,15 +117,10 @@ for root_lib_dir  in $ROOT_LIB_DIR  ; do
 				echo "    g++ $CFLAGS -g -o '$test_out_name' '$test_object_name' $LDFLAGS  -L'$out_dir'" >> $test_file
 				echo "    set -x" >> $test_file
 				echo "    $PREC '$test_out_name' '$PWD/small.txt' '$PWD/out.txt'" >> $test_file
-				echo "    $PREC '$test_out_name' '$PWD/medium.txt' '$PWD/out.txt'" >> $test_file
-				echo "    $PREC '$test_out_name' '$PWD/gpl.txt' '$PWD/out.txt'" >> $test_file
+				#echo "    $PREC '$test_out_name' '$PWD/medium.txt' '$PWD/out.txt'" >> $test_file
+				#echo "    $PREC '$test_out_name' '$PWD/gpl.txt' '$PWD/out.txt'" >> $test_file
 				echo "    set +x" >> $test_file
 				echo "}" >> $test_file
-				echo "" >> $test_file
-
-				echo "rm -rf '$PWD/out.txt'" >> $test_file
-				echo "rm -rf '$test_out_name'" >> $test_file
-				echo "rm -rf '$test_object_name'" >> $test_file
 				echo "" >> $test_file
 			fi
 			done
@@ -128,13 +132,15 @@ $PWD/roundup $test_file
 
 for root_lib_dir in $ROOT_LIB_DIR  ; do
 	make_dir="$root_lib_dir/$MF_SUFFIX/$LIBMAKE_SUFFIX/"
-	pushd $make_dir > /dev/null
-	make clean > /dev/null 2> /dev/null
-	rm -rf ./$LIBOUT_SUFFIX/
-	popd > /dev/null
+	if [ -d $make_dir ]; then
+		pushd $make_dir > /dev/null
+		make clean > /dev/null 2> /dev/null
+		rm -rf ./$LIBOUT_SUFFIX/
+		popd > /dev/null
+	fi
 done
 for clean_file in $TEST_BUILD; do
-	echo $clean_file
+	#echo $clean_file
 	rm -f $clean_file
 done
